+++ CHANGELOG.md
@@ -2,10 +2,27 @@
 
 ## [Unreleased]
 
+### Added
+- Adds Content-Type application/json to payload headers.
+- Adds getBody function to get the body from a message.
+- Adds getHtmlAndTextBody function to get the proper HtmlBody and TextBody fields.
+- Adds getMimePart function to get a mime part from a message.
+- Adds getTag function to get the tag from a message.
+- Adds getSubject function to get the subject from a message.
+- Adds tests to ensure that the json payload has the proper values.
+- Adds tests to ensure that the json payload has the proper HtmlBody and TextBody values.
+- Adds test to ensure that required fields are in the payload.
+
+### Updated
+- Refactor PostmarkTransportTest.
+- Payload does not contain empty keys.
+- Updates docblock getContact can receive a string or array.
+- Updates test for getting contacts into a comma seperated string.
+
 ## [2.1.3] - 2017-10-06
 
 ### Fixed
-- getAttachments should return array. [`9adaf6340f`](https://github.com/coconutcraig/laravel-postmark/commit/9adaf6340f) 
+- getAttachments should return array. [`9adaf6340f`](https://github.com/coconutcraig/laravel-postmark/commit/9adaf6340f)
 
 ## [2.1.2] - 2017-10-06